requirements.txt
tensorflow
keras
pandas==2.0.3
numpy==1.26.4
scikit-learn
matplotlib
Pillow
seaborn
streamlit